Make shorter bundle function

+++ index.ts
@@ -10,25 +10,27 @@ const {
   DBUSER = "root",
   DBPASS = "password",
   DBNAME = "anonymous",
+  GOOGLE_SERVICE_ACCOUNT,
 } = process.env
+
+if (!GOOGLE_SERVICE_ACCOUNT) {
+  console.error("Please provide a GOOGLE_SERVICE_ACCOUNT")
+  process.exit(1)
+}
 
 const locationDir = path.join(__dirname, "exports")
 
 const currentTime = dayjs().format("DD-MM-YYYY HH.mm.ss")
 const filename = `${locationDir}/${currentTime}`
-const filenameZip = `${filename}.zip`
 
 backupMysql(filename, DBHOST, parseInt(DBPORT), DBUSER, DBPASS, DBNAME).then(
   async (filenameSql) => {
     try {
-      await bundledToZip(filenameSql, filenameZip)
-      console.log("Success build this file")
-      fs.unlink(
-        filenameSql,
-        (err) => err && console.log("Can't delete temp data")
-      )
+      const fileZip = await bundledToZip(filenameSql)
+      console.log(`Success build to ${fileZip} file`)
     } catch (err) {
-      console.log("Can't Bundle this file", err)
+      console.log(`Can't Bundle ${filenameSql} file`, err)
+    } finally {
       fs.unlink(
         filenameSql,
         (err) => err && console.log("Can't delete temp data")
